requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=1.10.7
sqlalchemy>=2.0.9
web3>=6.2.0
python-dotenv>=1.0.0
httpx>=0.24.0
beautifulsoup4>=4.12.2
openai>=1.0.0
transformers>=4.28.1
python-multipart>=0.0.6
aiohttp>=3.8.4
torch>=2.0.0
tqdm>=4.65.0
eth-account>=0.8.1
